+++ tests/testthat/test-pluck.R
@@ -189,3 +189,60 @@ test_that("plucks by name", {
 test_that("can't pluck from complex", {
   expect_error( pluck( 1+2i, 1 ), "Don't know how to index object of type complex at level 1" )
 })
+
+
+# assign_in() ----------------------------------------------------------
+
+test_that("reduce_subset_call() type-checks", {
+  expect_identical(reduce_subset_call(NA, list(1, 4, "foo", 3)), quote(NA[[1]][[4]][["foo"]][[3]]))
+  expect_error(reduce_subset_call(NA, list(1, attr_getter("foo"), 3)), "must be names or positions, not a function")
+})
+
+test_that("assign_in() doesn't assign in the caller environment", {
+  x <- list(list(bar = 1, foo = 2))
+  assign_in(x, list(1, "foo"), value = 20)
+  expect_identical(x, list(list(bar = 1, foo = 2)))
+})
+
+test_that("assign_in() assigns", {
+  x <- list(list(bar = 1, foo = 2))
+  out <- assign_in(x, list(1, "foo"), value = 20)
+  expect_identical(out, list(list(bar = 1, foo = 20)))
+})
+
+test_that("pluck<- is an alias for assign_in()", {
+  x <- list(list(bar = 1, foo = 2))
+  pluck(x, 1, "foo") <- 30
+  expect_identical(x, list(list(bar = 1, foo = 30)))
+})
+
+test_that("assign_in() requires at least one location", {
+  x <- list("foo")
+  expect_error(assign_in(x, NULL, value = "foo"), "without pluck locations")
+  expect_error(pluck(x) <- "foo", "without pluck locations")
+})
+
+test_that("assign_in() requires existing location", {
+  x <- list(list(bar = 1, foo = 2))
+  expect_error(assign_in(x, 2, 10), "exceeds the length")
+  expect_error(assign_in(x, list(1, "baz"), 10), "Can't find name `baz`")
+})
+
+
+# modify_in() ----------------------------------------------------------
+
+test_that("modify_in() modifies in pluck location", {
+  x <- list(list(bar = 1, foo = 2))
+
+  out <- modify_in(x, list(1, "foo"), `+`, 100)
+  expect_identical(out, list(list(bar = 1, foo = 102)))
+
+  out <- modify_in(x, c(1, 1), `+`, 10)
+  expect_identical(out, list(list(bar = 11, foo = 2)))
+})
+
+test_that("modify_in() requires existing location", {
+  x <- list(list(bar = 1, foo = 2))
+  expect_error(modify_in(x, 2, `+`, 10), "exceeds the length")
+  expect_error(modify_in(x, list(1, "baz"), `+`, 10), "Can't find name `baz`")
+})